Refactor paging system for x86-64 architecture
- Transitioned from 32-bit to 64-bit paging with a 4-level page table structure.
- Updated kernel PML4 and page table management to support 2 MiB identity mapping.
- Enhanced page fault handling to accommodate 64-bit addresses.
- Modified memory management functions to return 64-bit addresses.
- Adjusted task scheduling and context switching to utilize 64-bit stack pointers.
- Updated assembly context switch code for x86-64 compatibility.
- Ensured proper handling of user and kernel page mappings with updated flags.
- Improved error handling and logging for memory allocation failures.

+++ include/pmm.h
@@ -7,12 +7,14 @@
 /* Page size: 4 KiB */
 #define PAGE_SIZE 4096
 
-/* Public API */
+/* Public API.
+ * Physical addresses are returned as uint64_t so they cast cleanly to
+ * pointers in 64-bit code; all RAM still lives below 4 GiB. */
 void      pmm_init(multiboot_info_t* mboot_info);
-uint32_t  pmm_alloc_page(void);    /* Returns physical address of a free 4K page, or 0 on failure */
-uint32_t  pmm_alloc_pages(uint32_t count);  /* Contiguous run of pages (for DMA/back buffers) */
-void      pmm_free_page(uint32_t addr);
-void      pmm_free_pages(uint32_t addr, uint32_t count);
+uint64_t  pmm_alloc_page(void);    /* Returns physical address of a free 4K page, or 0 on failure */
+uint64_t  pmm_alloc_pages(uint32_t count);  /* Contiguous run of pages (for DMA/back buffers) */
+void      pmm_free_page(uint64_t addr);
+void      pmm_free_pages(uint64_t addr, uint32_t count);
 void      pmm_dump_info(void);     /* Print stats to terminal */
 
 /* Statistics */
